# docs/swagger.yaml
definitions:
...
  dto.DocumentoVendaItemRequest:
    properties:
      created_by:
        type: integer
      cst_icms_id:
        type: integer
      cst_ipi_id:
        type: integer
      cst_pis_cofins_id:
        type: integer
      descricao_produto:
        type: string
      id:
        type: integer
      operacao_fiscal_id:
        type: integer
      percentual_desconto:
        type: number
      peso_bruto:
        type: number
      peso_liquido:
        type: number
      produto_id:
        type: integer
      quantidade:
        type: number
      updated_by:
        type: integer
      valor_desconto:
        type: number
      valor_frete:
        type: number
      valor_unitario:
        minimum: 0
        type: number
    required:
    - produto_id
    - quantidade
    type: object
  dto.DocumentoVendaItemResponse:
    properties:
      documento_venda_id:
        type: integer
      item:
        type: integer
      percentual_desconto:
        type: number
      peso_bruto:
        type: number
      peso_liquido:
        type: number
      produto_codigo:
        type: integer
      produto_id:
        type: integer
      produto_nome:
        type: string
      quantidade:
        type: number
      total_item:
        type: number
      total_item_com_frete:
        type: number
      valor_desconto:
        type: number
      valor_frete:
        type: number
      valor_unitario:
        type: number
    type: object
  dto.DocumentoVendaListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/dto.DocumentoVendaResponse'
        type: array
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
...
  dto.DocumentoVendaRequest:
    properties:
      condicao_pagamento_id:
        type: integer
      created_by:
        description: |-
          ============================================================
          AUDITORIA
          ============================================================
        type: integer
      data_previsao:
        type: string
      data_validade:
        description: |-
          ============================================================
          DATAS
          ============================================================
        type: string
      empresa_filial_id:
        description: |-
          ============================================================
          DADOS PRINCIPAIS
          ============================================================
        type: integer
      entidade_id:
        type: integer
      itens:
        description: |-
          ============================================================
          LISTAS
          ============================================================
        items:
          $ref: '#/definitions/dto.DocumentoVendaItemRequest'
        minItems: 1
        type: array
      numero:
        type: integer
      observacoes_adm:
        type: string
      observacoes_cliente:
        type: string
      observacoes_interna:
        description: |-
          ============================================================
          OBSERVAÇÕES
          ============================================================
        type: string
      pagamentos:
        items:
          $ref: '#/definitions/dto.DocumentoVendaPagamentoRequest'
        type: array
      tabela_preco_id:
        type: integer
      tipo_documento:
        description: 1-Orçamento, 2-Pedido
        type: integer
      tipo_entrega:
        description: RETIRADA, ENTREGA, LOCAL
        type: string
      tipo_operacao:
        description: 0-Entrada, 1-Saída
        type: integer
      transportadora_id:
        type: integer
      updated_by:
        type: integer
      valor_desconto:
        type: number
      valor_frete:
        description: |-
          ============================================================
          VALORES
          ============================================================
        type: number
      vendedor_id:
        description: |-
          ============================================================
          RELACIONAMENTOS
          ============================================================
        type: integer
    required:
    - condicao_pagamento_id
    - empresa_filial_id
    - tabela_preco_id
    - tipo_documento
    - tipo_operacao
    type: object
...
  utils.ErrorResponse:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /documentos-venda:
    get:
      consumes:
      - application/json
      description: Retorna uma lista paginada de documentos de venda, com suporte
        a filtros.
      parameters:
      - description: Número de registros por página
        in: query
        name: limit
        type: integer
      - description: Offset para a paginação
        in: query
        name: offset
        type: integer
      - description: Filtrar por tipo
        in: query
        name: tipo
        type: string
      - description: Filtrar por situação
        in: query
        name: situacao
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.DocumentoVendaListResponse'
        "500":
          description: Erro interno do servidor
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Lista os documentos de venda
      tags:
      - Documentos de Venda
    post:
      consumes:
      - application/json
      description: Cria um novo documento de venda com base nos dados fornecidos.
      parameters:
      - description: Dados para criar o documento de venda
        in: body
        name: documento
        required: true
        schema:
          $ref: '#/definitions/dto.DocumentoVendaRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.DocumentoVendaResponse'
        "400":
          description: Erro de validação ou dados inválidos
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Erro interno do servidor
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Cria um novo documento de venda
      tags:
      - Documentos de Venda
  /documentos-venda/{documento_id}:
    delete:
      consumes:
      - application/json
      description: Realiza a exclusão lógica de um documento de venda.
      parameters:
      - description: ID do Documento de Venda
        in: path
        name: documento_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: Nenhum conteúdo
        "400":
          description: Erro ao excluir
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "404":
          description: Documento de venda não encontrado
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Exclui um documento de venda
      tags:
      - Documentos de Venda
    get:
      consumes:
      - application/json
      description: Retorna os detalhes de um documento de venda específico.
      parameters:
      - description: ID do Documento de Venda
        in: path
        name: documento_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.DocumentoVendaResponse'
        "404":
          description: Documento de venda não encontrado
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Busca um documento de venda por ID
      tags:
      - Documentos de Venda
    put:
      consumes:
      - application/json
      description: Atualiza os dados de um documento de venda existente.
      parameters:
      - description: ID do Documento de Venda
        in: path
        name: documento_id
        required: true
        type: integer
      - description: Dados para atualizar o documento de venda
        in: body
        name: documento
        required: true
        schema:
          $ref: '#/definitions/dto.DocumentoVendaRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.DocumentoVendaResponse'
        "400":
          description: Erro de validação ou dados inválidos
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "404":
          description: Documento de venda não encontrado
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Atualiza um documento de venda
      tags:
      - Documentos de Venda
  /documentos-venda/{documento_id}/cancelar:
    post:
      consumes:
      - application/json
      description: Cancela um documento de venda, alterando seu status para cancelado.
      parameters:
      - description: ID do Documento de Venda
        in: path
        name: documento_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.DocumentoVendaResponse'
        "400":
          description: Erro ao cancelar documento
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "404":
          description: Documento de venda não encontrado
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "422":
          description: Documento não pode ser cancelado (validação de negócio)
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Cancela um documento de venda
      tags:
      - Documentos de Venda
  /documentos-venda/{documento_id}/emitir:
    post:
      consumes:
      - application/json
      description: Emite um documento de venda, alterando seu status para emitido.
      parameters:
      - description: ID do Documento de Venda
        in: path
        name: documento_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.DocumentoVendaResponse'
        "400":
          description: Erro ao emitir documento
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "404":
          description: Documento de venda não encontrado
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "422":
          description: Documento não pode ser emitido (validação de negócio)
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Emite um documento de venda
      tags:
      - Documentos de Venda
  /documentos-venda/{documento_id}/itens:
    get:
      consumes:
      - application/json
      description: Retorna todos os itens de um documento de venda específico.
      parameters:
      - description: ID do Documento de Venda
        in: path
        name: documento_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.DocumentoVendaItemResponse'
            type: array
        "404":
          description: Documento de venda não encontrado
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Lista os itens de um documento de venda
      tags:
      - Documentos de Venda
    post:
      consumes:
      - application/json
      description: Adiciona um novo item a um documento de venda existente.
      parameters:
      - description: ID do Documento de Venda
        in: path
        name: documento_id
        required: true
        type: integer
      - description: Dados do item
        in: body
        name: item
        required: true
        schema:
          $ref: '#/definitions/dto.DocumentoVendaItemRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.DocumentoVendaResponse'
        "400":
          description: Erro de validação ou dados inválidos
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "404":
          description: Documento de venda não encontrado
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "422":
          description: 'Erro de validação de negócio (ex: estoque insuficiente)'
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Adiciona um item ao documento de venda
      tags:
      - Documentos de Venda
  /documentos-venda/{documento_id}/itens/{item}:
    delete:
      consumes:
      - application/json
      description: Remove um item específico de um documento de venda.
      parameters:
      - description: ID do Documento de Venda
        in: path
        name: documento_id
        required: true
        type: integer
      - description: ID do Item
        in: path
        name: item
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.DocumentoVendaResponse'
        "400":
          description: Erro ao remover item
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "404":
          description: Documento de venda ou item não encontrado
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "422":
          description: Erro de validação de negócio
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Remove um item do documento de venda
      tags:
      - Documentos de Venda
    put:
      consumes:
      - application/json
      description: Atualiza os dados de um item específico de um documento de venda.
      parameters:
      - description: ID do Documento de Venda
        in: path
        name: documento_id
        required: true
        type: integer
      - description: ID do Item
        in: path
        name: item
        required: true
        type: integer
      - description: Dados atualizados do item
        in: body
        name: item
        required: true
        schema:
          $ref: '#/definitions/dto.DocumentoVendaItemRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.DocumentoVendaResponse'
        "400":
          description: Erro de validação ou dados inválidos
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "404":
          description: Documento de venda ou item não encontrado
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "422":
          description: Erro de validação de negócio
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Edita um item do documento de venda
      tags:
      - Documentos de Venda
  /documentos-venda/{documento_id}/pagamentos:
    get:
      consumes:
      - application/json
      description: Retorna todos os pagamentos de um documento de venda específico.
      parameters:
      - description: ID do Documento de Venda
        in: path
        name: documento_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.DocumentoVendaPagamentoResponse'
            type: array
        "404":
          description: Documento de venda não encontrado
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Lista os pagamentos de um documento de venda
      tags:
      - Documentos de Venda
    post:
      consumes:
      - application/json
      description: Adiciona um novo pagamento a um documento de venda existente.
      parameters:
      - description: ID do Documento de Venda
        in: path
        name: documento_id
        required: true
        type: integer
      - description: Dados do pagamento
        in: body
        name: pagamento
        required: true
        schema:
          $ref: '#/definitions/dto.DocumentoVendaPagamentoRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.DocumentoVendaResponse'
        "400":
          description: Erro de validação ou dados inválidos
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "404":
          description: Documento de venda não encontrado
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "422":
          description: 'Erro de validação de negócio (ex: valor excede o total)'
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Adiciona um pagamento ao documento de venda
      tags:
      - Documentos de Venda
  /documentos-venda/{documento_id}/pagamentos/{pagamento}:
    put:
      consumes:
      - application/json
      description: Atualiza os dados de um pagamento específico de um documento de
        venda.
      parameters:
      - description: ID do Documento de Venda
        in: path
        name: documento_id
        required: true
        type: integer
      - description: ID do Pagamento
        in: path
        name: pagamento
        required: true
        type: integer
      - description: Dados atualizados do pagamento
        in: body
        name: pagamento
        required: true
        schema:
          $ref: '#/definitions/dto.DocumentoVendaPagamentoRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.DocumentoVendaResponse'
        "400":
          description: Erro de validação ou dados inválidos
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "404":
          description: Documento de venda ou pagamento não encontrado
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "422":
          description: Erro de validação de negócio
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Edita um pagamento do documento de venda
      tags:
      - Documentos de Venda
...
swagger: "2.0"